SSCSCI-2519 Move isBenefitTypeChildSupportOrUc to SscsUtil
