+++ src/main/java/uk/gov/hmcts/reform/sscs/util/SscsUtil.java
@@ -118,6 +118,10 @@ public class SscsUtil {
             .orElse(null);
     }
 
+    public static boolean isBenefitTypeChildSupportOrUc(SscsCaseData sscsCaseData) {
+        return sscsCaseData.isBenefitType(Benefit.CHILD_SUPPORT) || sscsCaseData.isBenefitType(Benefit.UC);
+    }
+
     public static boolean isValidCaseState(State state, List<State> allowedStates) {
         return allowedStates.contains(state);
     }
