+++ src/test/java/uk/gov/hmcts/reform/sscs/util/SscsUtilTest.java
@@ -33,6 +33,7 @@ import static uk.gov.hmcts.reform.sscs.util.SscsUtil.getPostHearingReviewDocumen
 import static uk.gov.hmcts.reform.sscs.util.SscsUtil.getSscsType;
 import static uk.gov.hmcts.reform.sscs.util.SscsUtil.getWriteFinalDecisionDocumentType;
 import static uk.gov.hmcts.reform.sscs.util.SscsUtil.handleIbcaCase;
+import static uk.gov.hmcts.reform.sscs.util.SscsUtil.isBenefitTypeChildSupportOrUc;
 import static uk.gov.hmcts.reform.sscs.util.SscsUtil.updateHearingChannel;
 import static uk.gov.hmcts.reform.sscs.util.SscsUtil.updateHearingInterpreter;
 
@@ -55,6 +56,7 @@ import uk.gov.hmcts.reform.sscs.ccd.domain.AdjournCaseTypeOfHearing;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Adjournment;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appeal;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Appellant;
+import uk.gov.hmcts.reform.sscs.ccd.domain.Benefit;
 import uk.gov.hmcts.reform.sscs.ccd.domain.BenefitType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.CollectionItem;
 import uk.gov.hmcts.reform.sscs.ccd.domain.Correction;
@@ -117,6 +119,30 @@ class SscsUtilTest {
                 .isCorrectionFinalDecisionInProgress(YesNo.NO).build()).build();
         caseData = new SscsCaseData();
         caseData.setPostHearing(postHearing);
+    }
+
+    @Test
+    void givenChildSupportBenefit_isBenefitTypeChildSupportOrUcReturnsTrue() {
+        SscsCaseData sscsCaseData = SscsCaseData.builder()
+            .appeal(Appeal.builder().benefitType(BenefitType.builder().code(Benefit.CHILD_SUPPORT.getShortName()).build()).build())
+            .build();
+        assertTrue(isBenefitTypeChildSupportOrUc(sscsCaseData));
+    }
+
+    @Test
+    void givenUcBenefit_isBenefitTypeChildSupportOrUcReturnsTrue() {
+        SscsCaseData sscsCaseData = SscsCaseData.builder()
+            .appeal(Appeal.builder().benefitType(BenefitType.builder().code(Benefit.UC.getShortName()).build()).build())
+            .build();
+        assertTrue(isBenefitTypeChildSupportOrUc(sscsCaseData));
+    }
+
+    @Test
+    void givenOtherBenefit_isBenefitTypeChildSupportOrUcReturnsFalse() {
+        SscsCaseData sscsCaseData = SscsCaseData.builder()
+            .appeal(Appeal.builder().benefitType(BenefitType.builder().code(Benefit.PIP.getShortName()).build()).build())
+            .build();
+        assertFalse(isBenefitTypeChildSupportOrUc(sscsCaseData));
     }
 
     @Test
